+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## 0.3.1 (...)
 - use `autofocus` element on the previously selected IdP in the WAYF, so enter
   can be used to continue to that IdP immediately
+- translate language code to human readable language for display in UI
 
 ## 0.3.0 (2020-02-12)
 - small API update for `XmlIdpInfoSource` now taking array in contructor to